\name{Mean}
\alias{mean0}
\alias{skew0}
\alias{sd0}
\title{Mean, SD, skewness with na.omit}
\usage{
  mean0(v)
  sd0(v)
  skew0(v)
}
\description{
  \code{mean0} mean(v,na.rm=TRUE)

  \code{sd0} sd(v,na.rm=TRUE) 

  \code{skew0} moments::skewnewss(v,na.rm=TRUE)
}
\arguments{
  \item{v}{a numeric vector.}
}
